14. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A task

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	int a = 1, b = 2, s;
+	long int a = 1, b = 2, s;
 	int i;
 
 	printf("%ld, %ld", a, b);
